refactor: move ABICC compat layer into abicheck/compat/ subpackage
Structure:
  abicheck/compat/
    __init__.py         — re-exports CompatDescriptor, parse_descriptor
    descriptor.py       — ABICC XML descriptor parsing (from compat.py)
    xml_report.py       — ABICC XML report generation (from xml_report.py)
    abicc_dump_import.py — Perl ABI.dump importer (from abicc_dump_import.py)
    cli.py              — compat/compat-dump commands + all ABICC helpers

  abicheck/cli.py       — now only core dump/compare commands (~210 lines)

Backward-compat shims preserved:
  abicheck/compat.py           — re-exports CompatDescriptor, parse_descriptor
  abicheck/xml_report.py       — re-exports from compat.xml_report
  abicheck/abicc_dump_import.py — re-exports from compat.abicc_dump_import
  abicheck/cli.py              — re-exports all private compat helpers

Tests updated:
  test_abicc_dump_import.py  — import from abicheck.compat.abicc_dump_import
  test_cli_phase1.py         — monkeypatch targets → abicheck.compat.cli.*
  test_cli_unit.py           — monkeypatch targets → abicheck.compat.cli.*
  test_format_compliance.py  — source path → abicheck/compat/cli.py

Local: ruff clean, 1084 passed, 3 skipped, 0 failed
